requirements.txt
flask
gunicorn
python-dotenv
azure-cognitiveservices-speech
